backend/requirements.txt
Flask==2.1.1
Werkzeug==2.0.3
requests==2.27.1
python-dotenv==0.19.2
flask-cors==3.0.10
